+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [v1.7.0](http://github.com/ndustrialio/contxt-sdk-js/tree/v1.7.0) (2019-07-03)
+
+**Added**
+
+- Added `Coordinator.consent#getForCurrentApplication` for getting the current application version's consent form. The current access_token will be used to derive which application is being consented to.
+- Added `Coordinator.consent#accept` for accepting user consent to an applications
+
 ## [v1.6.0](http://github.com/ndustrialio/contxt-sdk-js/tree/v1.5.0) (2019-06-25)
 
 **Added**